tools/despatcher/pyproject.toml
[project]
name = "despatcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Shri",email = "me@drone-ah.com"}
]
readme = "README.md"
requires-python = "<3.14,>=3.8"
dependencies = [
    "python-frontmatter (>=1.1.0,<2.0.0)",
    "atproto (>=0.0.61,<0.0.62)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
